+++ cluster-operator/src/main/java/io/strimzi/operator/cluster/operator/resource/rolling/RollClientImpl.java
@@ -8,10 +8,8 @@ import io.strimzi.operator.cluster.model.NodeRef;
 import io.strimzi.operator.cluster.operator.resource.KafkaAgentClient;
 import io.strimzi.operator.cluster.operator.resource.KafkaBrokerConfigurationDiff;
 import io.strimzi.operator.cluster.operator.resource.KafkaBrokerLoggingConfigurationDiff;
-import io.strimzi.operator.common.Reconciliation;
 import io.strimzi.operator.common.UncheckedExecutionException;
 import io.strimzi.operator.common.UncheckedInterruptedException;
-import io.strimzi.operator.common.operator.resource.PodOperator;
 import org.apache.kafka.clients.admin.Admin;
 import org.apache.kafka.clients.admin.Config;
 import org.apache.kafka.clients.admin.DescribeClusterResult;
@@ -37,11 +35,9 @@ import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.ExecutionException;
 import java.util.stream.Collectors;
 
-class RollClientImpl implements RollClient, PlatformClient {
+class RollClientImpl implements RollClient {
 
     private final static int ADMIN_BATCH_SIZE = 200;
-    private final String namespace;
-    private final Reconciliation reconciliation;
 
     /** Return a future that completes when all of the given futures complete */
     @SuppressWarnings("rawtypes")
@@ -66,24 +62,11 @@ class RollClientImpl implements RollClient, PlatformClient {
 
     private final Admin admin;
 
-    private final PodOperator podOps;
-
     private final KafkaAgentClient kafkaAgentClient;
 
-    RollClientImpl(Reconciliation reconciliation,
-                   PodOperator podOps,
-                   String namespace,
-                   Admin admin, KafkaAgentClient kafkaAgentClient) {
-        this.reconciliation = reconciliation;
-        this.podOps = podOps;
-        this.namespace = namespace;
+    RollClientImpl(Admin admin, KafkaAgentClient kafkaAgentClient) {
         this.admin = admin;
         this.kafkaAgentClient = kafkaAgentClient;
-    }
-
-    @Override
-    public boolean isNotReady(NodeRef nodeRef) {
-        return !podOps.isReady(namespace, nodeRef.podName());
     }
 
     @Override
@@ -92,11 +75,6 @@ class RollClientImpl implements RollClient, PlatformClient {
         return BrokerState.fromValue((byte) kafkaAgentClient.getBrokerState(podName).code());
     }
 
-    @Override
-    public void restartNode(NodeRef nodeRef) {
-        var pod = podOps.get(namespace, nodeRef.podName());
-        podOps.restart(reconciliation, pod, 60_000);
-    }
 
     @Override
     public Collection<TopicListing> listTopics() {
